tshig mdzod chen mo: Fix tsheg drag entry
Fix the misparsed missing shad.  The two dots here are visarga (rnam
bcad), not gter tsheg.  The "sra" is actually s? - sa with halanta
(srog med):, but I'm not sure how to input this correctly.  Again, the
"random" collection of Tibetan syllables at the end are just plucked
out of the otherwise missing Chinese part of the entry and should
probably be g/c'ed.