49-LokeshChandraSkt: fix "rgu ba"
Should be "prāgbhāra" and the "g" is missing.  Compare with Negi.  The
scan at the archive.org has rather bad print quality in this
particular place, but I think you can see that ga in the sandhi if you
squint just right.

https://archive.org/details/dictionarytibetansanskritdictionarylokeshchandra_78_F/page/n511/mode/2up

Probably needs fixing in the source file too, but I'm not sure how to do that properly.